Move to new module name genesis-dev-tools
