+++ setup.cfg
@@ -1,11 +1,11 @@
 [metadata]
-name = Genesis Images
-summary = Configuraitons and build tools for Genesis images.
+name = Genesis Dev Tools
+summary = Tools to manager life cycle of Genesis projects.
 description_file =
     README.md
 author = Genesis Corporation
 author_email = anton.kremenetsky@gmail.com
-home_page = https://github.com/infraguys/genesis_images
+home_page = https://github.com/infraguys/genesis_dev_tools
 classifier =
     Intended Audience :: Developers
     License :: OSI Approved :: Apache Software License
@@ -18,9 +18,10 @@ classifier =
     Programming Language :: Python :: 3.12
     Programming Language :: Python :: 3.13
 
+
 [files]
 packages =
-    gcl_images
+    genesis_dev_tools
 
 
 [package_data]
@@ -30,4 +31,4 @@ ubuntu_24 =
 
 [entry_points]
 console_scripts =
-    genesis = gcl_images.cmd.cli:main
+    genesis = genesis_dev_tools.cmd.cli:main
