+++ tox.ini
@@ -11,7 +11,7 @@ skip_missing_interpreters = true
 deps = -r{toxinidir}/requirements.txt
        -r{toxinidir}/test-requirements.txt
 setenv =
-  PACKAGE_NAME=gcl_images
+  PACKAGE_NAME=genesis_dev_tools
   TEST_PATH={env:PACKAGE_NAME}/tests/unit
 commands =
   coverage run -p -m pytest {posargs} --timer-top-n=10 {env:TEST_PATH}
